+++ packages/frontend-framework/src/components/content/contentWithBar.tsx
@@ -16,24 +16,6 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-/**
- * Copyright © 2025 IAV GmbH Ingenieurgesellschaft Auto und Verkehr, All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- * SPDX-License-Identifier: Apache-2.0
- */
-
 import React, {useContext, useMemo} from "react";
 import "../css/globalColors.css";
 import {
@@ -43,7 +25,11 @@ import {
 } from "./contentBar";
 import {BasicContentbarWrapper} from "./basicContentbarWrapper";
 import {CustomContentbarWrapper} from "./customContentbarWrapper";
-import {ContentLayout, ContentLayoutAndStyleProps} from "./contentLayout";
+import {
+  ContentLayout,
+  ContentLayoutAndStyleProps,
+  LayoutBehaviour,
+} from "./contentLayout";
 import {ColorSettingsContext} from "../../contexts/colorsettings";
 
 export type ContentWithBarProps = {
@@ -67,12 +53,6 @@ export const ContentWithBar = (
   const contentAreaBackground =
     colorSettingsContext.currentColors.contentArea.backgroundColor;
 
-  const selectedContentWrapper = useMemo(() => {
-    return props.contentWrappers.find(
-      (currentWrapper) => currentWrapper.getId() === props.selectedId,
-    );
-  }, [props.contentWrappers, props.selectedId]);
-
   const contentBarStyles = useMemo(() => {
     const tempContentbarStyles: ContentBarStylesArray = [];
     Object.values(ContentBarStyles).forEach((contentBarStyle) => {
@@ -85,6 +65,18 @@ export const ContentWithBar = (
     });
     return tempContentbarStyles;
   }, [props.contentStyle]);
+
+  const getDisplayForSelected = (layoutBehaviour: LayoutBehaviour) => {
+    switch (layoutBehaviour) {
+      case LayoutBehaviour.FLEX:
+      case LayoutBehaviour.FLEX_COL:
+        return "flex";
+      case LayoutBehaviour.GRID:
+        return "flex";
+      default:
+        return "block";
+    }
+  };
 
   return (
     <div
@@ -119,7 +111,32 @@ export const ContentWithBar = (
           layoutBehaviour={props.layoutBehaviour}
           contentStyle={props.contentStyle}
         >
-          {selectedContentWrapper?.getContentAreaElement()}
+          {props.contentWrappers.map((wrapper) => {
+            const isSelected = wrapper.getId() === props.selectedId;
+
+            const displayStyle = isSelected
+              ? getDisplayForSelected(
+                  props.layoutBehaviour ?? LayoutBehaviour.NONE,
+                )
+              : "none";
+
+            const style: React.CSSProperties = {
+              display: displayStyle,
+              flexGrow: 1,
+              height: "100%",
+              width: "100%",
+              overflow: "auto",
+              ...(props.layoutBehaviour === LayoutBehaviour.FLEX_COL
+                ? {flexDirection: "column"}
+                : {}),
+            };
+
+            return (
+              <div key={wrapper.getId()} style={style}>
+                {wrapper.getContentAreaElement()}
+              </div>
+            );
+          })}
         </ContentLayout>
       </div>
     </div>